fix(api): convert newsletter/subscribe to ts
- ensure newsletter/subscribe return a response
